tests/wpt/tests/content-security-policy/style-src/style-blocked.html

<!DOCTYPE html>
<html>
<head>
    <meta http-equiv="Content-Security-Policy" content="style-src 'none'">
    <title>style-blocked</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script>
      async_test(t => {
        window.addEventListener('securitypolicyviolation',
                                t.step_func_done(e => {
          assert_equals(e.violatedDirective, "style-src-elem");
        }));
      }, "Violated directive is script-src-elem.");
    </script>
    <link rel="stylesheet" href="resources/blue.css">
</head>
<body>
  <script>
    test(t => {
      assert_equals(document.styleSheets.length, 1);
    }, "document.styleSheets should contain an item for the blocked CSS.");
  </script>
</body>
</html>
