LayoutTests/imported/w3c/web-platform-tests/content-security-policy/style-src/style-blocked-expected.txt


PASS Violated directive is script-src-elem.
FAIL document.styleSheets should contain an item for the blocked CSS. assert_equals: expected 1 but got 0
